Move nav layer to thumb left/right symlayers
Using combos made it hard to do shift+up/down.  So adding one extra
keypress to access the nav layer, but should make this easier in
general to access.  Skipping the conditional layers as I don't see a
way to have them one shot.  With this I can roll to each side to
access the nav layer.  Or at least in theory, I haven't tested this
yet.